requirements.txt
flask
python-dotenv
openai
flask-cors
python-telegram-bot
requests

